+++ src/gpu/cache/kernels/batch_norm_use_global_stats_gpu.cl
@@ -20,6 +20,6 @@
 KERNEL (batch_norm_use_global_stats_gpu)(const __global UNIT_TYPE* input, __global UNIT_TYPE* output, __global UNIT_TYPE* mean, __global UNIT_TYPE* variance)
 {
     const uint feature_id = get_global_id(1);
-    const uint linear_id = get_global_id(0) + feature_id * INPUT_BATCH_NUM + get_global_id(2) * INPUT_BATCH_NUM * INPUT_FEATURE_NUM;
+    const uint linear_id = (uint)get_global_id(0) + feature_id * INPUT_BATCH_NUM + (uint)get_global_id(2) * INPUT_BATCH_NUM * INPUT_FEATURE_NUM;
     output[linear_id] = (input[linear_id] - mean[feature_id]) / (sqrt(variance[feature_id]) + EPSILON);
 }
