+++ src/gpu/cache/kernels/convolution_gpu_yxfb_oiyx.naive.cl
@@ -38,8 +38,8 @@ KERNEL(convolution_gpu_yxfb_oiyx)(
     float result = bias[ofm_offset];
 
     bool finish = false;
-    const int out_x = get_global_id(1) - OUTPUT_PADDING_LOWER_SIZE_X;
-    const int out_y = get_global_id(2) - OUTPUT_PADDING_LOWER_SIZE_Y;
+    const int out_x = (uint)get_global_id(1) - OUTPUT_PADDING_LOWER_SIZE_X;
+    const int out_y = (uint)get_global_id(2) - OUTPUT_PADDING_LOWER_SIZE_Y;
 
     finish = out_x >= OUTPUT_SIZE_X || out_x < 0;
     finish = (out_y >= OUTPUT_SIZE_Y || out_y < 0) ? true : finish;
