+++ src/gpu/cache/kernels/convolution_gpu_yxfb_yxio_b16_fp16.cl
@@ -46,7 +46,7 @@ KERNEL(convolution_gpu_yxfb_yxio_b16_fp16)(
 
 
     const uint linear_id_xy = get_global_id(1) + get_global_size(1) * get_global_id(2);
-    uint global_id = ((get_global_id(0) / WORK_ITEMS_PER_SINGLE_BATCHES_ELEMENTS) + (linear_id_xy * FILTER_ARRAY_NUM + split_idx) * (FILTER_OUTPUT_FEATURE_NUM / OFM_PER_WORK_ITEM)) * WORK_ITEMS_PER_SINGLE_BATCHES_ELEMENTS;
+    uint global_id = (((uint)get_global_id(0) / WORK_ITEMS_PER_SINGLE_BATCHES_ELEMENTS) + (linear_id_xy * FILTER_ARRAY_NUM + split_idx) * (FILTER_OUTPUT_FEATURE_NUM / OFM_PER_WORK_ITEM)) * WORK_ITEMS_PER_SINGLE_BATCHES_ELEMENTS;
 
     const uint sub_group_id = get_local_id(0);
 
@@ -56,7 +56,7 @@ KERNEL(convolution_gpu_yxfb_yxio_b16_fp16)(
     const uint chunk_size = 1;
 #endif
 
-    const uint out_batch_id = chunk_size * sub_group_id + LOCAL_WORK_GROUP_SIZE * BATCHES_PER_WORK_ITEM * (get_group_id(0) % LOCAL_WORK_GROUPS_PER_SINGLE_BATCHES_ELEMENTS);
+    const uint out_batch_id = chunk_size * sub_group_id + LOCAL_WORK_GROUP_SIZE * BATCHES_PER_WORK_ITEM * ((uint)get_group_id(0) % LOCAL_WORK_GROUPS_PER_SINGLE_BATCHES_ELEMENTS);
     const uint out_x = get_global_id(1);
     const uint out_y = get_global_id(2);
