+++ src/gpu/cache/kernels/convolution_gpu_yxfb_yxio_b8.optimized.cl
@@ -32,7 +32,7 @@ KERNEL(convolution_gpu_yxfb_yxio_b8)(
 
     const uint linear_id_xy = get_global_id(1) + get_global_size(1) * get_global_id(2);
     // we're computing 8 OUTPUT_FEATURE_MAP so we must divide by 8, but we got 8 batches, so no division is needed.
-    uint global_id = (get_global_id(0) / batch_num) * batch_num + (linear_id_xy * FILTER_ARRAY_NUM + split_idx) * (FILTER_OUTPUT_FEATURE_NUM / OFM_PER_WORK_ITEM) * batch_num;
+    uint global_id = ((uint)get_global_id(0) / batch_num) * batch_num + (linear_id_xy * FILTER_ARRAY_NUM + split_idx) * (FILTER_OUTPUT_FEATURE_NUM / OFM_PER_WORK_ITEM) * batch_num;
 
     const uint out_batch_id = get_local_id(0);
     const uint out_x = get_global_id(1);
