+++ src/gpu/cache/kernels/scale_gpu.cl
@@ -30,24 +30,24 @@ KERNEL (scale_gpu)(const __global UNIT_TYPE* input, __global UNIT_TYPE* output, 
     const uint output_buffer_size_x = OUTPUT_PADDING_LOWER_SIZE_X + OUTPUT_SIZE_X + OUTPUT_PADDING_UPPER_SIZE_X;
     const uint output_buffer_size_y = OUTPUT_PADDING_LOWER_SIZE_Y + OUTPUT_SIZE_Y + OUTPUT_PADDING_UPPER_SIZE_Y;
 
-    const uint x = (get_global_id(2) % INPUT_SIZE_X);
-    const uint y = (get_global_id(2) / INPUT_SIZE_X);
+    const uint x = ((uint)get_global_id(2) % INPUT_SIZE_X);
+    const uint y = ((uint)get_global_id(2) / INPUT_SIZE_X);
 
     if ((x >= OUTPUT_SIZE_X) || (y >= OUTPUT_SIZE_Y))
         return;
 
 #if INPUT_BFYX_USED
-    const uint input_linear_id = x + INPUT_PADDING_LOWER_SIZE_X + input_buffer_size_x * (INPUT_PADDING_LOWER_SIZE_Y + y + input_buffer_size_y * (get_global_id(1) + get_global_id(0) * INPUT_FEATURE_NUM));
-    const uint output_linear_id = x + OUTPUT_PADDING_LOWER_SIZE_X + output_buffer_size_x * (OUTPUT_PADDING_LOWER_SIZE_Y + y + output_buffer_size_y * (get_global_id(1) + get_global_id(0) * OUTPUT_FEATURE_NUM));
+    const uint input_linear_id = x + INPUT_PADDING_LOWER_SIZE_X + input_buffer_size_x * (INPUT_PADDING_LOWER_SIZE_Y + y + input_buffer_size_y * ((uint)get_global_id(1) + (uint)get_global_id(0) * INPUT_FEATURE_NUM));
+    const uint output_linear_id = x + OUTPUT_PADDING_LOWER_SIZE_X + output_buffer_size_x * (OUTPUT_PADDING_LOWER_SIZE_Y + y + output_buffer_size_y * ((uint)get_global_id(1) + (uint)get_global_id(0) * OUTPUT_FEATURE_NUM));
 #else
-    const uint input_linear_id = get_global_id(0) + INPUT_BATCH_NUM * (get_global_id(1) + INPUT_FEATURE_NUM * (x + INPUT_PADDING_LOWER_SIZE_X + input_buffer_size_x * (INPUT_PADDING_LOWER_SIZE_Y + y)));
-    const uint output_linear_id = get_global_id(0) + OUTPUT_BATCH_NUM * (get_global_id(1) + OUTPUT_FEATURE_NUM * (x + OUTPUT_PADDING_LOWER_SIZE_X + output_buffer_size_x * (OUTPUT_PADDING_LOWER_SIZE_Y + y)));
+    const uint input_linear_id = (uint)get_global_id(0) + INPUT_BATCH_NUM * ((uint)get_global_id(1) + INPUT_FEATURE_NUM * (x + INPUT_PADDING_LOWER_SIZE_X + (uint)input_buffer_size_x * (INPUT_PADDING_LOWER_SIZE_Y + y)));
+    const uint output_linear_id = (uint)get_global_id(0) + OUTPUT_BATCH_NUM * ((uint)get_global_id(1) + OUTPUT_FEATURE_NUM * (x + OUTPUT_PADDING_LOWER_SIZE_X + (uint)output_buffer_size_x * (OUTPUT_PADDING_LOWER_SIZE_Y + y)));
 #endif
 
     const uint scale_batch_id = (SCALE_BATCH_NUM == 1) ? 0 : get_global_id(0);
     const uint scale_feature_id = (SCALE_FEATURE_NUM == 1) ? 0 : get_global_id(1);
-    const uint scale_x = (SCALE_SIZE_X == 1) ? 0 : ((SCALE_SIZE_Y == 1) ? (get_global_id(2) % INPUT_SIZE_X) : (get_global_id(2) % SCALE_SIZE_X));
-    const uint scale_y = (SCALE_SIZE_Y == 1) ? 0 : ((SCALE_SIZE_X == 1) ? (get_global_id(2) / INPUT_SIZE_X) : (get_global_id(2) / SCALE_SIZE_X));
+    const uint scale_x = (SCALE_SIZE_X == 1) ? 0 : ((SCALE_SIZE_Y == 1) ? ((uint)get_global_id(2) % INPUT_SIZE_X) : ((uint)get_global_id(2) % SCALE_SIZE_X));
+    const uint scale_y = (SCALE_SIZE_Y == 1) ? 0 : ((SCALE_SIZE_X == 1) ? ((uint)get_global_id(2) / INPUT_SIZE_X) : ((uint)get_global_id(2) / SCALE_SIZE_X));
 #if SCALE_BFYX_USED
     const uint scale_linear_id = scale_x + SCALE_SIZE_X * (scale_y + SCALE_SIZE_Y * (scale_feature_id + scale_batch_id * SCALE_FEATURE_NUM));
 #else
